Authentication_Microservice/Authentication/pyproject.toml
[project]
name = "Dodgygeezers"
version = "0.1.0"
description = "Micro-service for user auth and dashboard for Dodgygeezers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "debugpy>=1.8.9",
    "fastapi[standard]>=0.115.5",
    "grpcio-tools>=1.68.1",
    "grpcio>=1.68.1",
    "itsdangerous>=2.2.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.6.1",
    "python-jose>=3.3.0",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.32.1",
]

[tool.uv]
dev-dependencies = [
    "pytest<8.0.0,>=7.4.3",
    "mypy<2.0.0,>=1.8.0",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
]
